auth

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,12 +1,22 @@
 import React from "react";
-import { useWorkoutsContext } from '../hooks/useWorkoutContext'
+import { useWorkoutsContext } from '../hooks/useWorkoutContext';
+import { useAuthContext } from "../hooks/useAuthContext";
+
 
 const WorkoutDetails = ({ workout }) => {
+  const {user} = useAuthContext();
   const { dispatch } = useWorkoutsContext();
 
   const handleClick = async () => {
+    if(!user){
+      return
+    }
+
     const response = await fetch('/api/workouts/' + workout._id, {
       method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
 
     const data = await response.json()
